requirements.txt
flask~=2.2.2
flask-smorest
python-dotenv
sqlalchemy~=1.4.42
flask-sqlalchemy
marshmallow~=3.18.0
flask_jwt_extended
passlib~=1.7.4
flask-migrate
flask-cors
pytest~=7.2.0
coverage
Flask-APScheduler
requests
asyncpraw
asyncio
psycopg2
gunicorn